+++ test/integration/events/merge_events.js
@@ -13,12 +13,6 @@ describe(require('../../__fixtures/utils/test_helper').create().testName(__filen
   var publisherclient;
   var listenerclient;
 
-  /*
-   This test demonstrates starting up the happn service -
-   the authentication service will use authTokenSecret to encrypt web tokens identifying
-   the logon session. The utils setting will set the system to log non priority information
-   */
-
   after(function (done) {
     this.timeout(10000);