+++ test/integration/security/admin-login-encrypted-payloads.js
@@ -117,7 +117,7 @@ describe(require('../../__fixtures/utils/test_helper').create().testName(__filen
 
                 client.on('/test/path', function(data){
 
-                  expect(data["test"]).to.be("data");
+                  expect(data.test).to.be("data");
 
                   client.disconnect({reconnect:false});