+++ test/unit/data/b_data_functional_sanity.js
@@ -372,7 +372,7 @@ describe(require('../../__fixtures/utils/test_helper').create().testName(__filen
           }
         };
 
-        listCriteria.criteria.$not['_id'] = {
+        listCriteria.criteria.$not._id = {
           $regex: new RegExp(".*_notok_.*")
         };